added unit tests for mean, median, and sd
