+++ src/print_fires.py
